# Splits g() into 2 inlined funcs.