Eliminates calls & bound checking. 

benchmark            old ns/op     new ns/op     delta
BenchmarkWrite-4     368095        212280        -42.33%
BenchmarkChunk-4     17496         10271         -41.30%
BenchmarkXOF-4       10.5          5.93          -43.52%

benchmark            old MB/s     new MB/s     speedup
BenchmarkWrite-4     89.02        154.36       1.73x
BenchmarkXOF-4       95.48        168.68       1.77x