@lukechampine 2020-01-10
Wonderful, thank you! I knew there was some way to improve the inlining.

There was a (trivial) merge conflict, so I rebased and merged locally.